+++ src/padkit/hash.c
@@ -6,7 +6,7 @@
 #include "padkit/debug.h"
 #include "padkit/hash.h"
 
-uint_fast64_t hash_str(char const* restrict str, size_t register n) {
+uint_fast64_t hash_str(char const* restrict str, uint64_t register n) {
     DEBUG_ERROR_IF(str == NULL)
     uint_fast64_t register hash = 5381;
     while (n-- && *str) hash += (hash << 5) + (uint_fast64_t)(*(str++));
